Rename max_..._actions.py
