+++ pixi.lock
@@ -4847,8 +4847,8 @@ packages:
   timestamp: 1719463874284
 - pypi: .
   name: lcm
-  version: 0.1.dev199+g06f86d2.d20250314
-  sha256: d3941d0c1373efb631cd6e527b750ccd1df8a5587dfc266c297c159f656cec03
+  version: 0.1.dev201+gfae6a04
+  sha256: 0eec480760d770196b75da15155cbc17c100ebfd8c46beb844ecb7d9ef9e8e19
   requires_dist:
   - dags
   - numpy
