+++ pyproject.toml
@@ -196,6 +196,8 @@ extend-ignore = [
 "docs/source/conf.py" = ["E501", "ERA001", "DTZ005"]
 "tests/*" = ["PLR2004", "D101"]
 "examples/*" = ["INP001", "D101"]
+"src/lcm/max_Qc_over_d" = ["N999"]
+"src/lcm/max_Q_over_c" = ["N999"]
 "explanations/*" = ["INP001", "B018", "T201", "E402", "PD008"]
 "scripts/*" = ["INP001", "D101", "RET503"]
 "**/*.ipynb" = ["FBT003", "E402", "D101"]
