+++ src/lcm/max_Q_over_c.py
