+++ src/lcm/max_Qc_over_d.py
